Silence -Werror,-Wimplicit-int-float-conversion
> vcl/source/window/window2.cxx:588:20: error: implicit conversion from 'long' to 'double' changes value from -9223372036854775807 to -9223372036854775808 [-Werror,-Wimplicit-int-float-conversion]
>         if ( nN == -LONG_MAX )
>                 ~~ ^~~~~~~~~

etc. with Clang 10

+++ vcl/source/window/window2.cxx
@@ -585,9 +585,9 @@ static void lcl_HandleScrollHelper( ScrollBar* pScrl, double nN, bool isMultiply
     {
         long nNewPos = pScrl->GetThumbPos();
 
-        if ( nN == -LONG_MAX )
+        if ( nN == double(-LONG_MAX) )
             nNewPos += pScrl->GetPageSize();
-        else if ( nN == LONG_MAX )
+        else if ( nN == double(LONG_MAX) )
             nNewPos -= pScrl->GetPageSize();
         else
         {
@@ -657,9 +657,9 @@ bool Window::HandleScrollCommand( const CommandEvent& rCmd,
                         if ( nScrollLines == COMMAND_WHEEL_PAGESCROLL )
                         {
                             if ( pData->GetDelta() < 0 )
-                                nLines = -LONG_MAX;
+                                nLines = double(-LONG_MAX);
                             else
-                                nLines = LONG_MAX;
+                                nLines = double(LONG_MAX);
                         }
                         else
                             nLines = pData->GetNotchDelta() * nScrollLines;
